requirements.txt
Django==3.0.3
djangorestframework==3.12.2
django-rest-swagger==2.2.0
drf-yasg==1.20.0
django-cors-headers==3.7.0